Remove unnecessary `stream()`

+++ modules/models/src/main/kotlin/org/cafejojo/schaapi/models/NodeSequenceUtil.kt
@@ -26,8 +26,8 @@ class NodeSequenceUtil<N : Node> {
      * @param minimumCount the minimum number of times a node needs to occur
      * @return the set of nodes occurring at least [minimumCount] times
      */
-    fun findFrequentNodesInSequences(sequences: Collection<List<N>>, minimumCount: Int): Map<N, Long> {
-        val nodeCounts = CustomEqualsHashMap<N, Long>(Node.Companion::equiv, Node::equivHashCode)
+    fun findFrequentNodesInSequences(sequences: Collection<List<N>>, minimumCount: Int): Map<N, Int> {
+        val nodeCounts = CustomEqualsHashMap<N, Int>(Node.Companion::equiv, Node::equivHashCode)
         val sequenceSets = sequences.map { sequence ->
             CustomEqualsHashSet<N>(Node.Companion::equiv, Node::equivHashCode).also { it.addAll(sequence) }
         }
